requirements.txt
click==6.6
coreapi==2.1.1
coreapi-cli==1.0.6
dj-database-url==0.4.1
Django==1.10.4
djangorestframework==3.5.3
gunicorn==19.6.0
itypes==1.1.0
Pygments==2.1.3
requests==2.12.4
simplejson==3.10.0
twitter==1.17.1
uritemplate==3.0.0
whitenoise==3.2.2
